requirements.txt
fastapi 
uvicorn 
psycopg2-binary 
python-dotenv 
paho-mqtt
firebase-admin